refactor: introduce `proto_reader.rs` (#1624)
