+++ src/config/reader.rs
@@ -1,10 +1,7 @@
-use std::collections::{HashMap, VecDeque};
 use std::path::Path;
 use std::sync::Arc;
 
-use anyhow::Context;
-use prost_reflect::prost_types::{FileDescriptorProto, FileDescriptorSet};
-use protox::file::{FileResolver, GoogleFileResolver};
+use anyhow::{anyhow, Context};
 use rustls_pemfile;
 use rustls_pki_types::{
     CertificateDer, PrivateKeyDer, PrivatePkcs1KeyDer, PrivatePkcs8KeyDer, PrivateSec1KeyDer,
@@ -13,23 +10,25 @@ use rustls_pki_types::{
 use super::{ConfigModule, Content, Link, LinkType};
 use crate::config::{Config, ConfigReaderContext, Source};
 use crate::merge_right::MergeRight;
+use crate::proto_reader::ProtoReader;
 use crate::resource_reader::ResourceReader;
 use crate::rest::EndpointSet;
 use crate::runtime::TargetRuntime;
-use crate::valid::{Valid, Validator};
 
 /// Reads the configuration from a file or from an HTTP URL and resolves all
 /// linked extensions to create a ConfigModule.
 pub struct ConfigReader {
     runtime: TargetRuntime,
     resource_reader: ResourceReader,
+    proto_reader: ProtoReader,
 }
 
 impl ConfigReader {
     pub fn init(runtime: TargetRuntime) -> Self {
         Self {
             runtime: runtime.clone(),
-            resource_reader: ResourceReader::init(runtime),
+            resource_reader: ResourceReader::init(runtime.clone()),
+            proto_reader: ProtoReader::init(runtime),
         }
     }
 
@@ -57,10 +56,7 @@ impl ConfigReader {
             return Ok(config_module);
         }
 
-        for i in 0..links.len() {
-            let config_link = links
-                .get(i)
-                .context(format!("Expected a link at index: {i} but found none"))?;
+        for config_link in links.iter() {
             let path = Self::resolve_path(&config_link.src, parent_dir);
 
             let source = self.resource_reader.read_file(&path).await?;
@@ -84,28 +80,17 @@ impl ConfigReader {
                     }
                 }
                 LinkType::Protobuf => {
-                    let parent_descriptor = self.read_proto(&source.path).await?;
-
-                    let id = Valid::from_option(
-                        parent_descriptor.package.clone(),
-                        format!(
-                            "Package name is not defined for proto file: {:?} with link id: {:?}",
-                            parent_descriptor.name, config_link.id
-                        ),
-                    )
-                    .trace(&format!("link[{}]", i))
-                    .trace("schema")
-                    .to_result()?;
-
-                    let mut descriptors = self.resolve_descriptors(parent_descriptor).await?;
-                    let mut file_descriptor_set = FileDescriptorSet::default();
-
-                    file_descriptor_set.file.append(&mut descriptors);
-
-                    config_module
-                        .extensions
-                        .grpc_file_descriptors
-                        .push(Content { id: Some(id), content: file_descriptor_set });
+                    let path = Self::resolve_path(&config_link.src, parent_dir);
+                    let meta = self.proto_reader.read(path).await?;
+                    if let Some(id) = meta.package.clone() {
+                        config_module
+                            .extensions
+                            .grpc_file_descriptors
+                            .push(Content { id: Some(id), content: meta.descriptor_set.clone() });
+                    } else {
+                        return Err(anyhow!("Package name is not defined for proto file"))
+                            .context(config_link.src.to_string());
+                    }
                 }
                 LinkType::Script => {
                     config_module.extensions.script = Some(content);
@@ -233,45 +218,6 @@ impl ConfigReader {
         Ok(config_module)
     }
 
-    /// Performs BFS to import all nested proto files
-    async fn resolve_descriptors(
-        &self,
-        parent_proto: FileDescriptorProto,
-    ) -> anyhow::Result<Vec<FileDescriptorProto>> {
-        let mut descriptors: HashMap<String, FileDescriptorProto> = HashMap::new();
-        let mut queue = VecDeque::new();
-        queue.push_back(parent_proto.clone());
-
-        while let Some(file) = queue.pop_front() {
-            for import in file.dependency.iter() {
-                let proto = self.read_proto(import).await?;
-                if descriptors.get(import).is_none() {
-                    queue.push_back(proto.clone());
-                    descriptors.insert(import.clone(), proto);
-                }
-            }
-        }
-        let mut descriptors = descriptors
-            .into_values()
-            .collect::<Vec<FileDescriptorProto>>();
-        descriptors.push(parent_proto);
-        Ok(descriptors)
-    }
-
-    /// Tries to load well-known google proto files and if not found uses normal
-    /// file and http IO to resolve them
-    async fn read_proto(&self, path: &str) -> anyhow::Result<FileDescriptorProto> {
-        let content = if let Ok(file) = GoogleFileResolver::new().open_file(path) {
-            file.source()
-                .context("Unable to extract content of google well-known proto file")?
-                .to_string()
-        } else {
-            self.resource_reader.read_file(path).await?.content
-        };
-
-        Ok(protox_parse::parse(path, &content)?)
-    }
-
     /// Checks if path is absolute else it joins file path with relative dir
     /// path
     fn resolve_path(src: &str, root_dir: Option<&Path>) -> String {
@@ -286,24 +232,11 @@ impl ConfigReader {
 
 #[cfg(test)]
 mod test_proto_config {
-    use std::collections::VecDeque;
-    use std::path::{Path, PathBuf};
+    use std::path::PathBuf;
 
-    use anyhow::{Context, Result};
-    use pretty_assertions::assert_eq;
+    use anyhow::Result;
 
     use crate::config::reader::ConfigReader;
-    use crate::valid::ValidationError;
-
-    #[tokio::test]
-    async fn test_resolve() {
-        // Skipping IO tests as they are covered in reader.rs
-        let reader = ConfigReader::init(crate::runtime::test::init(None));
-        reader
-            .read_proto("google/protobuf/empty.proto")
-            .await
-            .unwrap();
-    }
 
     #[tokio::test]
     async fn test_proto_no_pkg() -> Result<()> {
@@ -312,88 +245,8 @@ mod test_proto_config {
         let mut proto_no_pkg = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
         proto_no_pkg.push("src/grpc/tests/proto_no_pkg.graphql");
         let config_module = reader.read(proto_no_pkg.to_str().unwrap()).await;
-        let validation = config_module
-            .err()
-            .unwrap()
-            .downcast::<ValidationError<String>>()?;
-        proto_no_pkg.pop();
-        proto_no_pkg.push("proto");
-        proto_no_pkg.push("news_no_pkg.proto");
-        let err = &validation.as_vec().first().unwrap().message;
-        let trace = &validation.as_vec().first().unwrap().trace;
-        assert!(err.starts_with("Package name is not defined for proto file"));
-
-        let expected_trace = ["schema", "link[0]"]
-            .iter()
-            .map(|s| s.to_string())
-            .collect::<VecDeque<String>>();
-        assert_eq!(&expected_trace, trace);
+        assert!(config_module.is_err());
         Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_nested_imports() -> Result<()> {
-        let root_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-        let mut test_dir = root_dir.join(file!());
-        test_dir.pop(); // config
-        test_dir.pop(); // src
-
-        let mut root = test_dir.clone();
-        root.pop();
-
-        test_dir.push("grpc"); // grpc
-        test_dir.push("tests"); // tests
-        test_dir.push("proto"); // proto
-
-        let mut test_file = test_dir.clone();
-
-        test_file.push("nested0.proto"); // nested0.proto
-        assert!(test_file.exists());
-        let test_file = test_file.to_str().unwrap().to_string();
-
-        let runtime = crate::runtime::test::init(None);
-        let file_rt = runtime.file.clone();
-
-        let reader = ConfigReader::init(runtime);
-        let helper_map = reader
-            .resolve_descriptors(reader.read_proto(&test_file).await?)
-            .await?;
-        let files = test_dir.read_dir()?;
-        for file in files {
-            let file = file?;
-            let path = file.path();
-            let path_str =
-                path_to_file_name(path.as_path()).context("It must be able to extract path")?;
-            let source = file_rt.read(&path_str).await?;
-            let expected = protox_parse::parse(&path_str, &source)?;
-            let actual = helper_map
-                .iter()
-                .find(|v| v.package.eq(&expected.package))
-                .unwrap();
-
-            assert_eq!(&expected.dependency, &actual.dependency);
-        }
-
-        Ok(())
-    }
-
-    fn path_to_file_name(path: &Path) -> Option<String> {
-        let components: Vec<_> = path.components().collect();
-
-        // Find the index of the "src" component
-        if let Some(src_index) = components.iter().position(|&c| c.as_os_str() == "src") {
-            // Reconstruct the path from the "src" component onwards
-            let after_src_components = &components[src_index..];
-            let result = after_src_components
-                .iter()
-                .fold(PathBuf::new(), |mut acc, comp| {
-                    acc.push(comp);
-                    acc
-                });
-            Some(result.to_str().unwrap().to_string())
-        } else {
-            None
-        }
     }
 }
